requirements.txt
numpy
pyqt5
pyqtgraph
PyOpenGL
numpy-stl
